WC-2981: index.js: Do write _after_ processing transaction
In the test suites, we're using a dummy port that operates
synchronously.

This fails when we're testing debug mode because this routine is
responsible for creating the `request` and `response` properties on the
transaction.

So instead, move the write call.  This still passes tests, and I'd
expect it to make bugger all difference in production.  (If
`_startTimeout` is taking long periods of time, then you have other
problems.)