WC-2988: index.js: Make ModbusRTU an EventEmitter.
`node-red-modbus-contrib` monitors for disconnection events by hooking
the `close` event from the port by accessing the `_port` property of the
connection object.

This seems like a bad thing to do, but there are numerous cases where a
user of this library would need to receive such events.  So, better
plan: make the client itself an `EventEmitter` so it can proxy such
events.