WC-2988: index.js: Prevent memory leak when calling `.open`.
If a port gets closed, and we call `.open`, a new copy of the `data`
handler is created and added to the port's `data` handler list, which
will race the first and cause double-handling.

Thus, we need to remove the old listener first before we add a new one.
HOWEVER, since it's a new copy, we don't have the reference to the old
one to hand to `removeListener`, so a better plan is needed:

1. move the logic out to its own private function.
2. we need a consistent reference, so use the `.bind` function once in
   our constructor, so we can re-use that same reference when calling
   `removeListener`
3. pass our premade reference to `_port.on`

This will save memory and prevent an ugly double-up situation.